Reorganize files as part of the workflow directory
