+++ src/workflows/recipes/components/ApprovalDetails.js
